+++ stm32f7508_freertos-fp/src/main/java/com/microej/example/partial/support/PartialDisplay.java
@@ -1,5 +1,5 @@
-/**
- * Copyright 2020 MicroEJ Corp. All rights reserved.
+/*
+ * Copyright 2020-2021 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
@@ -59,9 +59,10 @@ public class PartialDisplay extends Display {
 
 	@Override
 	public void flush(MicroUIGraphicsContext gc, Image backBuffer, int x, int y, int width, int height) {
-		this.frameBuffer.drawImage(backBuffer, x, y, this.flushWidth, this.flushHeight,
-				x + this.flushX, y + this.flushY, this.flushWidth, this.flushHeight);
-
+		int flushWidth = Math.min(width, this.flushWidth);
+		int flushHeight = Math.min(height, this.flushHeight);
+		this.frameBuffer.drawImage(backBuffer, x, y, flushWidth, flushHeight, x + this.flushX, y + this.flushY,
+				flushWidth, flushHeight);
 		repaint();
 	}
 }