+++ com.microej.example.partial.support/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.1 - 2021-03-23
+
+### Fixed
+
+- Fix incorrect flush area in frontpanel.
+- Fix render policy issue when rendering the desktop while the root widget does not fill it.
+- Fix render policy issue when requesting to render a zone bigger than the widget.
+
 ## 1.0.0 - 2021-02-09
 
 ### Added